// src/enum/validation-type.ts
export enum ValidationType {
    
    SRV_SIDE_EXC = "SRV_SIDE_EXC",
    ATTACHMENT_KEY_NOT_FOUND = "ATTACHMENT_KEY_NOT_FOUND",
    USER_NOT_FOUND = "USER_NOT_FOUND",
    INVALID_EMAIL_PASSWORD = "INVALID_EMAIL_PASSWORD",
    USER_ALREADY_EXIST = "USER_ALREADY_EXIST",
    USER_REGISTERED_SUCCESSFULLY = "USER_REGISTERED_SUCCESSFULLY",
    PRODUCT_NOT_FOUND = "PRODUCT_NOT_FOUND",
    PRODUCT_ID_NOT_EXIST = "PRODUCT_ID_NOT_EXIST",
    INVALID_OPERATION = "INVALID_OPERATION",
    CUSTOMER_ID_NOT_EXIST = "CUSTOMER_ID_NOT_EXIST",
    CUSTOMER_NOT_FOUND = "CUSTOMER_NOT_FOUND",
    INVALID_USER = "INVALID_USER",
    NOT_SUFFCIENT_STOCK = "NOT_SUFFCIENT_STOCK",
    ORDER_NOT_FOUND = "ORDER_NOT_FOUND",
    ALREADY_PAID_THIS_ORDER = "ALREADY_PAID_THIS_ORDER",
    ALREADY_CANCEL_THIS_ORDER = "ALREADY_CANCEL_THIS_ORDER",
    USER_ID_NOT_EXIST="USER_ID_NOT_EXIST",
    USER_MANAGE_SUCCESSFULLY="USER_MANAGE_SUCCESSFULLY",
    INVALID_DATE="INVALID_DATE"
    
}